client/requirements.txt
# =====================================================
# ЗАВИСИМОСТИ ДЛЯ КЛИЕНТА (macOS) - ОБНОВЛЕНО
# =====================================================

# =====================================================
# gRPC КЛИЕНТ
# =====================================================
grpcio
grpcio-tools
protobuf

# =====================================================
# РАСПОЗНАВАНИЕ РЕЧИ (STT)
# =====================================================
SpeechRecognition
sounddevice        # ← ЗАМЕНА PyAudio на sounddevice

# =====================================================
# ЗАХВАТ ЭКРАНА
# =====================================================
Pillow
mss               # ← ТОЛЬКО mss: Быстрый захват экрана

# =====================================================
# ОБРАБОТКА АУДИО
# =====================================================
numpy
pydub
# FLAC поддержка (встроен в Speech Recognition 1.5.0)
# Speech Recognition автоматически использует встроенный FLAC для оптимизации

# =====================================================
# СИСТЕМНЫЕ ЗАВИСИМОСТИ (устанавливаются через Homebrew)
# =====================================================
# brew install switchaudio-osx  # Переключение аудио устройств
# brew install sparkle          # Система автообновлений

# =====================================================
# УТИЛИТЫ
# =====================================================
rich
pynput
PyYAML            # ← Для чтения конфигурации
pystray

# =====================================================
# PyObjC - ВСЕ НЕОБХОДИМЫЕ ФРЕЙМВОРКИ ДЛЯ macOS
# =====================================================
pyobjc-core
pyobjc-framework-CoreAudio
pyobjc-framework-CoreFoundation
pyobjc-framework-AVFoundation
pyobjc-framework-Quartz
pyobjc-framework-ApplicationServices
pyobjc-framework-SystemConfiguration

# =====================================================
# UI КОМПОНЕНТЫ
# =====================================================
rumps             # ← macOS tray приложение

# =====================================================
# СЕТЕВЫЕ ОПЕРАЦИИ
# =====================================================
aiohttp           # ← Для HTTP запросов и скачивания обновлений

# =====================================================
# ДОПОЛНИТЕЛЬНЫЕ УТИЛИТЫ
# =====================================================
pathlib2          # ← Для совместимости с Python 3.6+
psutil
